backend/pyproject.toml
[tool.poetry]
name = "campuswap-backend"
version = "0.1.0"
description = "CampuSwap microservice backend"
authors = ["CampuSwap Team <campuswap@example.com>"]
readme = "README.md"
packages = [{ include = "apps" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = { extras = ["standard"], version = "^0.29.0" }
sqlalchemy = "^2.0.29"
alembic = "^1.13.1"
pydantic = "^2.6.4"
pydantic-settings = "^2.2.1"
httpx = "^0.27.0"
apscheduler = "^3.10.4"
python-jose = "^3.3.0"
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
email-validator = "^2.1.1"
scikit-learn = "^1.4.2"
numpy = "^1.26.4"
pandas = "^2.2.1"
pyjwt = "^2.8.0"
loguru = "^0.7.2"
psycopg = { extras = ["binary"], version = "^3.1.18" }
PyMySQL = "^1.1.0"
mysqlclient = "^2.2.4"

[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
flake8 = "^7.0.0"
mypy = "^1.8.0"
isort = "^5.13.2"
pytest = "^8.1.1"
pytest-asyncio = "^0.23.6"
httpx = "^0.27.0"
pre-commit = "^3.6.2"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
